feat: Mui Alert signalizing that lecture name is identical to lecture code
Added Alert signalizing that a lecture should be renamed for readability reasons.

Tooltip of the cog button hasn't been removed, hanging over cog button if lecture still hasn't been renamed,
will still state that the the name of the lecture is the same as lecture code.

To decide if we want to remove the tooltip.

Refs: #18