feat: Added loading state to asynchronous action buttons. Added new component GraderLoadingButton for that purpose.
fix: In last commit I forgot to add navigation in manual submission view to next submission, fixed now.
Back button in Edit Submission View now navigates back to submission from which it was accessed or to Grading
Table if it was accessed through clicking on edit icon button. Previously it was only navigating back to grading table.

Ref: #19